# backend/authentificate/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Account API on Go documentation",
        "title": "Account Service API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8081",
    "basePath": "/api/Accounts",
    "paths": {
        "/all_dates": {
            "get": {
                "description": "Получение всех дат из БД",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Dates"
                ],
                "summary": "Получение дат",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/enter_site": {
            "post": {
                "description": "Генерация и задание токена для каждого пользователя(случайный uuid)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Токен для пользователя",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/set_min_day/{minDay}": {
            "put": {
                "description": "Указатние какая дата в БД будет минимальной. Нужно для тестов",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Dates"
                ],
                "summary": "Указание минмального дня в базе данных",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Day",
                        "name": "minDay",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/signIn": {
            "post": {
                "description": "Вход в аккаунт и установка токенов",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Вход в аккаунт",
                "parameters": [
                    {
                        "description": "Данные для входа",
                        "name": "info",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.SignUpData"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "access_token",
                        "schema": {
                            "$ref": "#/definitions/models.TokenResponse"
                        }
                    },
                    "400": {
                        "description": "invalid request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "user not found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "cannot generate token",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/signUp": {
            "post": {
                "description": "Регистрация аккаунта и установка токенов",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Регистрация аккаунта",
                "parameters": [
                    {
                        "description": "Данные для входа",
                        "name": "info",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.SignUpData"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "access_token",
                        "schema": {
                            "$ref": "#/definitions/models.TokenResponse"
                        }
                    },
                    "400": {
                        "description": "invalid request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "cannot create account",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "models.SignUpData": {
            "type": "object",
            "properties": {
                "password": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                }
            }
        },
        "models.TokenResponse": {
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string"
                }
            }
        }
    }
}